% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Electric}
\alias{Electric}
\title{Electric}
\usage{
Electric(server, port = 80, username = NULL, password = NULL)
}
\arguments{
\item{server}{Server address}

\item{port}{Server port. Default is 80}

\item{username}{Username. Default is NULL}

\item{passowrd}{Password Default is NULL}
}
\description{
Created a Connection to an Electric-Dataland Server
}
\examples{
Electric("https://electric-dataland.org",3000)
}
\keyword{connection}
