% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweets.R
\name{getTweets}
\alias{getTweets}
\title{getTweets}
\usage{
getTweets(collection_title, collection_id = NULL)
}
\arguments{
\item{collection_title}{Title of the collection}

\item{collection_is}{Id of the collection. Default is NULL}
}
\description{
get all tweets for a collections
}
\examples{
getTweets("testcollection")
}
