% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivityBars.R
\name{plotActivityBars}
\alias{plotActivityBars}
\title{plotActivityBars}
\usage{
plotActivityBars(tweets, plotly = FALSE)
}
\arguments{
\item{tweets}{A `tweets` dataframe.}

\item{plotly}{Create interactive plot using plotly. Default is FALSE.}
}
\value{
A plot.
}
\description{
Creates a barplot that shows the frequency of tweets per day.
}
\examples{
plotActivityBars(tweets, plotly=TRUE)
}
\keyword{activity}
