% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivitySmooth.R
\name{plotActivitySmooth}
\alias{plotActivitySmooth}
\title{plotActivitySmooth
Creates a smooth lineplot that shows the frequency of tweets per day.}
\usage{
plotActivitySmooth(tweets, span = 0.1, plotly = FALSE)
}
\arguments{
\item{tweets}{A `tweets` dataframe.}

\item{span}{Defines the smoothness of the plot. Default is 0.1.}

\item{plotly}{Create interactive plot using plotly. Default is FALSE.}
}
\value{
A plot.
}
\description{
plotActivitySmooth
Creates a smooth lineplot that shows the frequency of tweets per day.
}
\examples{
plotActivitySmooth(tweets, span = 0.2, plotly=TRUE)
}
\keyword{activity}
