% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweets.R
\name{getTweets}
\alias{getTweets}
\title{getTweets}
\usage{
getTweets(collection_title, collection_id = NULL, start_date = "",
  end_date = "", chunksize = 10, profile = "vectR", debug = FALSE)
}
\arguments{
\item{collection_title}{Title of the collection}

\item{start_date}{Tweets created before this date "YYYY-MM-DD" are ignored. Default is ""}

\item{end_date}{Tweets created after this date "YYYY-MM-DD" are ignored. Default is ""}

\item{profile}{Default is vectR}

\item{collection_is}{Id of the collection. Default is NULL}

\item{limit}{max. number of tweets fetched}
}
\value{
A dataframe with following coloums

 `id` id of the tweet

 `screen_name` of the twitteruser

 `name` of the twitteruser

 `text` tweettext

 `created_at` date in raw twitter api format

 `created_at` date in Date() format

 `hashtags` hastags, separated by ,

 `user_mentions` screen_names of mentioned users, separated by ,

 `retweet_screen_name` screenname of the retweeted user, if the tweet is a retweet.

 Additional coloumns for annotated values
}
\description{
Get all tweets for a collection.
}
\examples{
getTweets("testcollection")
}
